Move GTFSCache functionality to an abstract class to allow GTFS API to share functionality, and use soft references.
